+++ example/lib/pages/dialogs_page.dart
@@ -248,26 +248,36 @@ class MacosuiSheet extends StatelessWidget {
             Row(
               mainAxisAlignment: MainAxisAlignment.center,
               children: [
-                MacosListTile(
-                  leading: const Icon(CupertinoIcons.lightbulb),
+                const MacosListTile(
+                  leading: Icon(CupertinoIcons.lightbulb),
                   title: Text(
                     'A robust library of Flutter components for macOS',
-                    style: MacosTheme.of(context).typography.headline,
+                    //style: MacosTheme.of(context).typography.headline,
                   ),
                   subtitle: Text(
                     'Create native looking macOS applications using Flutter',
-                    style:
-                        MacosTheme.of(context).typography.subheadline.copyWith(
-                              color: MacosColors.systemGrayColor,
-                            ),
                   ),
                 ),
+              ],
+            ),
+            const SizedBox(height: 16),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                const MacosListTile(
+                  leading: Icon(CupertinoIcons.bolt),
+                  title: Text(
+                    'Create beautiful macOS applications in minutes',
+                    //style: MacosTheme.of(context).typography.headline,
+                  ),
+                ),
+                const SizedBox(width: 10),
               ],
             ),
             const Spacer(),
             PushButton(
               buttonSize: ButtonSize.large,
-              child: const Text('Dismiss'),
+              child: const Text('Get started'),
               onPressed: () => Navigator.of(context).pop(),
             ),
             const SizedBox(height: 50),